+++ docs/conf.py
@@ -22,10 +22,10 @@ import sphinx_rtd_theme
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
-sys.path.insert(0, os.path.abspath('../djangocms_helper/test_utils'))
+sys.path.insert(0, os.path.abspath('../app_helper/test_utils'))
 sys.path.insert(1, os.path.abspath('..'))
-import cms_helper  # isort:skip
-cms_helper.setup_nocms()
+import helper  # isort:skip
+helper.setup_nocms()
 
 # -- General configuration ------------------------------------------------